mv ws_fetch kernelspecs and contents_manager fixtures to pytest_plugin for 3rd party usage
