+++ tests/services/contents/test_manager.py
@@ -15,14 +15,6 @@ from ...utils import expected_http_error
 
 # -------------- Functions ----------------------------
 
-# contents_manager_atomic = pytest.fixture(lambda tmp_path: FileContentsManager(root_dir=str(tmp_path), use_atomic_writing=True))
-# contents_manager_nonatomic = pytest.fixture(lambda tmp_path: FileContentsManager(root_dir=str(tmp_path), use_atomic_writing=False))
-
-@pytest.fixture(params=[True, False])
-def contents_manager(request, tmp_path):
-    return FileContentsManager(root_dir=str(tmp_path), use_atomic_writing=request.param)
-
-
 def _make_dir(contents_manager, api_path):
     """
     Make a directory.
